Fix incorrect distant player positions when attached
When a player is attached to an object (e.g. riding a cart, animal,
vehicle) and that object goes out of range, the player's position
gets messed up. With this change, if a player is attached to an
object, that object always sends its position to clients (if unlimited
range for players is set).